fixed checkin refactor
