+++ app/checkin.html
@@ -13,6 +13,7 @@
   <script src="https://www.gstatic.com/firebasejs/7.14.3/firebase-app.js"></script>
   <script src="https://www.gstatic.com/firebasejs/7.14.3/firebase-auth.js"></script>
   <script src="https://www.gstatic.com/firebasejs/7.14.3/firebase-firestore.js"></script>
+  <!--script type="text/javascript" src="./js/js_functions/checkin_functions.js"></script-->
 </head>
 
 <body>
@@ -42,7 +43,9 @@
   <button id="startFlowBtn" style="display: block;">Start My Flow</button>
   <button id="cancelBtn" style="display: block;">Cancel</button>
 
-  <script> require('./js/js_functions/checkin_functions.js') </script>
+  <!--script> require('./js/js_functions/checkin_functions.js') </script-->
+
+  <!--script type="text/javascript" src="./js/js_functions/checkin_functions.js"></script-->
   <script> require('./js/checkin.js') </script>
 
 </body>
